Add __dso_handle workaround to attempt to pass CI

+++ tests/python/pants_test/backend/python/tasks/test_ctypes_integration.py
@@ -90,4 +90,5 @@ class CTypesIntegrationTest(PantsRunIntegrationTest):
         self.assertIn('Text from 3rdparty!', pants_run.stdout_data)
         self.assertTrue(os.path.exists(cereal_outfile))
       finally:
-        os.remove(cereal_outfile)
+        if os.path.exists(cereal_outfile):
+          os.remove(cereal_outfile)
